backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
pandas==2.1.4
numpy==1.26.3
pillow==10.2.0
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
ultralytics==8.1.0
scikit-learn==1.3.2
python-dotenv==1.0.0
reportlab==4.0.9
reportlab==4.0.9
Pillow==10.2.0